solution-2 | method-1
- using a flag bool chan answeredInTime
- run two routines for timer and answer
- result based on which channel first writes into the flag
