--- students/ehernandez/p2.csv
@@ -1,4 +1,0 @@
-2x2,4
-5x6,30
-8x2,16
-9x6,54
