--- students/latentgenius/questions.csv
@@ -1,13 +1,0 @@
-5+5,10
-7+3,10
-8+3,11
-8+6,14
-1+1,2
-1+2,3
-3+1,4
-1+4,5
-2+3,5
-5+1,6
-3+3,6
-2+4,6
-5+2,7
